+++ packages/tooltip/src/components/tooltip/Tooltip.tsx
@@ -1,41 +1,12 @@
 import { SmallText } from "@stenajs-webui/core";
 import * as React from "react";
-import { Popover, PopoverProps } from "../popover/Popover";
-import { defaultTooltipTheme, TooltipTheme } from "./TooltipTheme";
+import { Popover, PopoverProps, PopoverVariant } from "../popover/Popover";
 
 export interface TooltipProps extends Omit<PopoverProps, "content"> {
-  type?: TooltipInfoType;
+  variant?: PopoverVariant;
   label: string;
-  theme?: TooltipTheme;
 }
 
-type TooltipInfoType = "info" | "warning" | "error";
-
-export const Tooltip: React.FC<TooltipProps> = ({
-  type = "info",
-  label,
-  theme = defaultTooltipTheme,
-  ...popoverProps
-}) => {
-  const textColor =
-    type === "error"
-      ? theme.textColorError
-      : type === "warning"
-      ? theme.textColorWarning
-      : theme.textColor;
-
-  const background =
-    type === "error"
-      ? theme.backgroundError
-      : type === "warning"
-      ? theme.backgroundWarning
-      : theme.background;
-
-  return (
-    <Popover
-      {...popoverProps}
-      background={background}
-      content={<SmallText color={textColor}>{label}</SmallText>}
-    />
-  );
+export const Tooltip: React.FC<TooltipProps> = ({ label, ...popoverProps }) => {
+  return <Popover {...popoverProps} content={<SmallText>{label}</SmallText>} />;
 };